Implemented the view page to use the list
Also implemented a getter that performed the time transformation on the
a milestone object.
Currently implementing the functionality to convert each time with
respect to the task that comes directly before it in the list, this will
then pass the transformed list back to the server and display it so that
the start times of each item will be listed not its duration.